Refactors, merge resource_packs & behavior_packs into only one folder, add description to the config and bump the API version
