+++ src/protocolsupportpocketstuff/packet/play/ResourcesPackInfoPacket.java
@@ -4,7 +4,7 @@ import io.netty.buffer.ByteBuf;
 import protocolsupport.api.Connection;
 import protocolsupport.protocol.typeremapper.pe.PEPacketIDs;
 import protocolsupportpocketstuff.packet.PEPacket;
-import protocolsupportpocketstuff.util.resourcepacks.ResourcePack;
+import protocolsupportpocketstuff.resourcepacks.ResourcePack;
 import protocolsupportpocketstuff.util.serializer.ResourcePackSerializer;
 
 import java.util.List;
