+++ src/protocolsupportpocketstuff/resourcepacks/InvalidResourcePackException.java
@@ -1,4 +1,4 @@
-package protocolsupportpocketstuff.util.resourcepacks;
+package protocolsupportpocketstuff.resourcepacks;
 
 public class InvalidResourcePackException extends RuntimeException {
 	public InvalidResourcePackException(String reason) {
